# openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: ILM Red API
  description: |
    Cloud-native API for digital knowledge management platform.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (JWT)**: Include the JWT in the `Authorization` header
       ```
       Authorization: Bearer eyJhbGciOiJSUzI1NiI...
       ```

    2. **API Key**: Include your API key in the `X-API-Key` header
       ```
       X-API-Key: ilm_live_abc123...
       ```

    ## Rate Limiting

    | Tier       | Requests/min | AI Tokens/day |
    |------------|--------------|---------------|
    | Free       | 60           | 10,000        |
    | Premium    | 300          | 100,000       |
    | Enterprise | 1,000        | Unlimited     |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset`

    ## Pagination

    List endpoints support pagination with `page` and `limit` query parameters.
    Response includes a `pagination` object with total count and page info.

  version: 1.0.0
  contact:
    name: ILM Red Support
    email: support@ilm-red.com
    url: https://docs.ilm-red.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.ilm-red.com/v1
    description: Production
  - url: https://api-staging.ilm-red.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email/password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/api-keys:
    get:
      tags:
        - Authentication
      summary: List API keys
      operationId: listApiKeys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
    post:
      tags:
        - Authentication
      summary: Create API key
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 100
                permissions:
                  type: array
                  items:
                    type: string
                expiresAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyWithSecret'
  /auth/api-keys/{keyId}:
    delete:
      tags:
        - Authentication
      summary: Revoke API key
      operationId: revokeApiKey
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: API key revoked

  # ============= BOOKS =============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            documentation:
              type: string
              format: uri
      example:
        error:
          code: "BOOK_NOT_FOUND"
          message: "The requested book does not exist or you don't have access"
          requestId: "req_abc123xyz789"
          timestamp: "2025-07-15T10:30:00Z"
          documentation: "https://api.ilm-red.com/docs/errors#BOOK_NOT_FOUND"
...
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Seconds until access token expires
        tokenType:
          type: string
          default: Bearer
      example:
        accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2MWUzZGI1Mi1mYWMxLTRjMDAtOGZiZi03NTI4N2VkODA4NGEiLCJlbWFpbCI6InVzZXJAZXhhbXBsZS5jb20iLCJpYXQiOjE3MjA1MzQ4MDB9.mock_signature"
        refreshToken: "rt_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
        expiresIn: 900
        tokenType: "Bearer"
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        keyPrefix:
          type: string
        permissions:
          type: array
          items:
            type: string
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ApiKeyWithSecret:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          properties:
            key:
              type: string
              description: Full API key (only shown once)
...